Make Pinot metrics pluggable
