+++ pinot-common/src/main/java/org/apache/pinot/common/metrics/PinotMetricUtils.java
@@ -18,82 +18,186 @@
  */
 package org.apache.pinot.common.metrics;
 
+import java.lang.reflect.Constructor;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
-import org.apache.pinot.common.exception.InvalidConfigException;
+import org.apache.pinot.plugin.metrics.yammer.YammerMetricsFactory;
+import org.apache.pinot.spi.annotations.metrics.MetricsFactory;
+import org.apache.pinot.spi.annotations.metrics.PinotMetricsFactory;
 import org.apache.pinot.spi.env.PinotConfiguration;
 import org.apache.pinot.spi.metrics.PinotGauge;
 import org.apache.pinot.spi.metrics.PinotJmxReporter;
+import org.apache.pinot.spi.metrics.PinotMeter;
 import org.apache.pinot.spi.metrics.PinotMetricName;
 import org.apache.pinot.spi.metrics.PinotMetricsRegistry;
-import org.apache.pinot.common.metrics.yammer.YammerGauge;
-import org.apache.pinot.common.metrics.yammer.YammerJmxReporter;
-import org.apache.pinot.common.metrics.yammer.YammerMetricName;
-import org.apache.pinot.common.metrics.yammer.YammerMetricsRegistry;
+import org.apache.pinot.spi.metrics.PinotTimer;
+import org.reflections.Reflections;
+import org.reflections.scanners.TypeAnnotationsScanner;
+import org.reflections.util.ClasspathHelper;
+import org.reflections.util.ConfigurationBuilder;
+import org.reflections.util.FilterBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 
 public class PinotMetricUtils {
   private static final Logger LOGGER = LoggerFactory.getLogger(PinotMetricUtils.class);
-  public static final String LIBRARY_NAME_KEY = "libraryName";
-  public static final String YAMMER_KEY = "yammer";
+  private static final String METRICS_PACKAGE_REGEX_PATTERN = ".*\\.plugin\\.metrics\\..*";
 
-  private static String LIBRARY_TO_USE = YAMMER_KEY;
+  private static PinotMetricsFactory _pinotMetricsFactory = null;
 
-  public static void init(PinotConfiguration metricsConfiguration)
-      throws InvalidConfigException {
-    String libraryName = metricsConfiguration.getProperty(PinotMetricUtils.LIBRARY_NAME_KEY);
-    if (libraryName == null) {
-      return;
+  private static final Map<PinotMetricsRegistry, Boolean> metricsRegistryMap = new ConcurrentHashMap<>();
+  private static final Map<MetricsRegistryRegistrationListener, Boolean> metricsRegistryRegistrationListenersMap =
+      new ConcurrentHashMap<>();
+
+  public static void init(PinotConfiguration metricsConfiguration) {
+    // Initializes PinotMetricsFactory.
+    initializePinotMetricsFactory(metricsConfiguration);
+
+    // Initializes metrics using the metrics configuration.
+    initializeMetrics(metricsConfiguration);
+    registerMetricsRegistry(getPinotMetricsRegistry());
+  }
+
+  /**
+   * Initializes PinotMetricsFactory with metrics configurations.
+   * @param metricsConfiguration The subset of the configuration containing the metrics-related keys
+   */
+  private static void initializePinotMetricsFactory(PinotConfiguration metricsConfiguration) {
+    Set<Class<?>> classes = getMetricsClasses();
+    for (Class<?> clazz : classes) {
+      MetricsFactory annotation = clazz.getAnnotation(MetricsFactory.class);
+      if (annotation.enabled()) {
+        try {
+          PinotMetricsFactory pinotMetricsFactory = (PinotMetricsFactory) clazz.newInstance();
+          pinotMetricsFactory.init(metricsConfiguration);
+          registerMetricsFactory(pinotMetricsFactory);
+        } catch (Exception e) {
+          LOGGER.error("Caught exception while initializing pinot metrics registry: {}, skipping it", clazz, e);
+        }
+      }
     }
-    switch (libraryName) {
-      case YAMMER_KEY:
-        LIBRARY_TO_USE = YAMMER_KEY;
-        break;
-      // TODO: support more libraries.
-      default:
-        throw new InvalidConfigException("PinotMetricsRegistry for " + libraryName + " cannot be initialized.");
+  }
+
+  private static Set<Class<?>> getMetricsClasses() {
+    Reflections reflections = new Reflections(
+        new ConfigurationBuilder().setUrls(ClasspathHelper.forPackage("org.apache.pinot"))
+            .filterInputsBy(new FilterBuilder.Include(METRICS_PACKAGE_REGEX_PATTERN))
+            .setScanners(new TypeAnnotationsScanner()));
+    return reflections.getTypesAnnotatedWith(MetricsFactory.class, true);
+  }
+
+  /**
+   * Initializes the metrics system by initializing the registry registration listeners present in the configuration.
+   *
+   * @param configuration The subset of the configuration containing the metrics-related keys
+   */
+  private static void initializeMetrics(PinotConfiguration configuration) {
+    synchronized (PinotMetricUtils.class) {
+      List<String> listenerClassNames = configuration.getProperty("metricsRegistryRegistrationListeners",
+          Arrays.asList(JmxReporterMetricsRegistryRegistrationListener.class.getName()));
+
+      // Build each listener using their default constructor and add them
+      for (String listenerClassName : listenerClassNames) {
+        try {
+          Class<? extends MetricsRegistryRegistrationListener> clazz =
+              (Class<? extends MetricsRegistryRegistrationListener>) Class.forName(listenerClassName);
+          Constructor<? extends MetricsRegistryRegistrationListener> defaultConstructor =
+              clazz.getDeclaredConstructor();
+          MetricsRegistryRegistrationListener listener = defaultConstructor.newInstance();
+
+          LOGGER.info("Registering metricsRegistry to listener {}", listenerClassName);
+          addMetricsRegistryRegistrationListener(listener);
+        } catch (Exception e) {
+          LOGGER
+              .warn("Caught exception while initializing MetricsRegistryRegistrationListener " + listenerClassName, e);
+        }
+      }
     }
-    LOGGER.info("Setting metric library to: " + LIBRARY_TO_USE);
+    LOGGER.info("Number of listeners got registered: {}", metricsRegistryRegistrationListenersMap.size());
+  }
+
+  /**
+   * Adds a metrics registry registration listener. When adding a metrics registry registration listener, events are
+   * fired to add all previously registered metrics registries to the newly added metrics registry registration
+   * listener.
+   *
+   * @param listener The listener to add
+   */
+  private static void addMetricsRegistryRegistrationListener(MetricsRegistryRegistrationListener listener) {
+    synchronized (PinotMetricUtils.class) {
+      metricsRegistryRegistrationListenersMap.put(listener, Boolean.TRUE);
+
+      // Fire events to register all previously registered metrics registries
+      Set<PinotMetricsRegistry> metricsRegistries = metricsRegistryMap.keySet();
+      LOGGER.info("Number of metrics registry: {}", metricsRegistries.size());
+      for (PinotMetricsRegistry metricsRegistry : metricsRegistries) {
+        listener.onMetricsRegistryRegistered(metricsRegistry);
+      }
+    }
+  }
+
+  /**
+   * Registers the metrics registry with the metrics helper.
+   *
+   * @param registry The registry to register
+   */
+  private static void registerMetricsRegistry(PinotMetricsRegistry registry) {
+    synchronized (PinotMetricUtils.class) {
+      metricsRegistryMap.put(registry, Boolean.TRUE);
+
+      // Fire event to all registered listeners
+      Set<MetricsRegistryRegistrationListener> metricsRegistryRegistrationListeners =
+          metricsRegistryRegistrationListenersMap.keySet();
+      for (MetricsRegistryRegistrationListener metricsRegistryRegistrationListener : metricsRegistryRegistrationListeners) {
+        metricsRegistryRegistrationListener.onMetricsRegistryRegistered(registry);
+      }
+    }
+  }
+
+  /**
+   * Registers an metrics factory.
+   */
+  private static void registerMetricsFactory(PinotMetricsFactory metricsFactory) {
+    LOGGER.info("Registering metrics factory: {}", metricsFactory.getMetricsFactoryName());
+    _pinotMetricsFactory = metricsFactory;
   }
 
   public static PinotMetricsRegistry getPinotMetricsRegistry() {
-    switch (LIBRARY_TO_USE) {
-      case YAMMER_KEY:
-        return new YammerMetricsRegistry();
-      //TODO: support more libraries.
-      default:
-        return new YammerMetricsRegistry();
+    if (_pinotMetricsFactory == null) {
+      // If no Pinot metrics factory gets registered, just initialize a Yammer metrics factory by default.
+      _pinotMetricsFactory = new YammerMetricsFactory();
     }
+    return _pinotMetricsFactory.getPinotMetricsRegistry();
   }
 
-  public static PinotMetricName generatePinotMetricName(Class<?> klass, String name) {
-    switch (LIBRARY_TO_USE) {
-      case YAMMER_KEY:
-        return new YammerMetricName(klass, name);
-      //TODO: support more libraries.
-      default:
-        return new YammerMetricName(klass, name);
-    }
+  public static PinotMetricName makePinotMetricName(Class<?> klass, String name) {
+    return _pinotMetricsFactory.makePinotMetricName(klass, name);
   }
 
-  public static <T> PinotGauge<T> generatePinotGauge(Function<Void, T> condition) {
-    switch (LIBRARY_TO_USE) {
-      case YAMMER_KEY:
-        return new YammerGauge<T>(condition);
-      //TODO: support more libraries.
-      default:
-        return new YammerGauge<T>(condition);
-    }
+  public static <T> PinotGauge<T> makePinotGauge(Function<Void, T> condition) {
+    return _pinotMetricsFactory.makePinotGauge(condition);
   }
 
-  public static PinotJmxReporter generatePinotJmxReporter(PinotMetricsRegistry metricsRegistry) {
-    switch (LIBRARY_TO_USE) {
-      case YAMMER_KEY:
-        return new YammerJmxReporter(metricsRegistry);
-      //TODO: support more libraries.
-      default:
-        return new YammerJmxReporter(metricsRegistry);
-    }
+  public static <T> PinotGauge<T> makeGauge(PinotMetricsRegistry registry, PinotMetricName name, PinotGauge<T> gauge) {
+    return registry.newGauge(name, gauge);
+  }
+
+  public static PinotTimer makePinotTimer(PinotMetricsRegistry registry, PinotMetricName name, TimeUnit durationUnit,
+      TimeUnit rateUnit) {
+    return registry.newTimer(name, durationUnit, rateUnit);
+  }
+
+  public static PinotMeter makePinotMeter(PinotMetricsRegistry registry, PinotMetricName name, String eventType, TimeUnit unit) {
+    return registry.newMeter(name, eventType, unit);
+  }
+
+  public static PinotJmxReporter makePinotJmxReporter(PinotMetricsRegistry metricsRegistry) {
+    return _pinotMetricsFactory.makePinotJmxReporter(metricsRegistry);
   }
 }
