+++ pinot-plugins/pinot-metrics/pinot-yammer/src/main/java/org/apache/pinot/plugin/metrics/yammer/YammerMetric.java
@@ -16,15 +16,21 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.metrics;
+package org.apache.pinot.plugin.metrics.yammer;
 
-import org.apache.pinot.spi.metrics.PinotMetricsRegistry;
+import com.yammer.metrics.core.Metric;
+import org.apache.pinot.spi.metrics.PinotMetric;
 
 
-/**
- * Interface to implement operations that occur whenever a new MetricsRegistry is registered with the PinotMetricUtils.
- *
- */
-public interface MetricsRegistryRegistrationListener {
-  void onMetricsRegistryRegistered(PinotMetricsRegistry metricsRegistry);
+public class YammerMetric implements PinotMetric {
+  private final Metric _metric;
+
+  public YammerMetric(Metric metric) {
+    _metric = metric;
+  }
+
+  @Override
+  public Metric getMetric() {
+    return _metric;
+  }
 }
