+++ pinot-spi/src/main/java/org/apache/pinot/spi/annotations/metrics/MetricsFactory.java
@@ -16,28 +16,24 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.metrics.yammer;
+package org.apache.pinot.spi.annotations.metrics;
 
-import com.yammer.metrics.core.Counter;
-import com.yammer.metrics.core.Metric;
-import org.apache.pinot.spi.metrics.PinotCounter;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 
-public class YammerCounter extends YammerMetric implements PinotCounter {
-  private final Counter _counter;
+/**
+ * Annotation for Pinot metrics.
+ *
+ * NOTE:
+ *   - The annotated class must implement the MinionEventObserverFactory interface
+ *   - The annotated class must be under the package of name 'org.apache.pinot.*.metrics.*' to be auto-registered.
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface MetricsFactory {
 
-  public YammerCounter(Counter counter) {
-    super(counter);
-    _counter = counter;
-  }
-
-  @Override
-  public Object getCounter() {
-    return _counter;
-  }
-
-  @Override
-  public Metric getMetric() {
-    return _counter;
-  }
+  boolean enabled() default true;
 }
